Small refactor to the runtime
